+++ src/daft-core/src/array/ops/mod.rs
@@ -13,7 +13,6 @@ mod comparison;
 mod concat;
 mod concat_agg;
 mod count;
-mod date;
 mod exp;
 mod filter;
 mod float;
@@ -46,6 +45,7 @@ mod struct_;
 mod sum;
 mod take;
 pub(crate) mod tensor;
+mod time;
 pub mod trigonometry;
 mod truncate;
 mod utf8;
